server/requirements.txt
Flask==0.10.1
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql==0.0.1
mysql-connector-python==2.0.4
MySQL-python==1.2.5
pyaml==15.8.2
PyYAML==3.11
Werkzeug==0.10.4
wheel==0.24.0
